add severity coloring to sensor/entity cards
