+++ src/common/style/severity_map.ts
@@ -1,0 +1,7 @@
+export const severityMap = {
+  red: "var(--label-badge-red)",
+  green: "var(--label-badge-green)",
+  yellow: "var(--label-badge-yellow)",
+  normal: "var(--label-badge-blue)",
+  undefined: "",
+};
